feat: introduce pytest
