+++ requirements.txt
@@ -16,6 +16,12 @@ attrs==25.3.0
     #   fourcipp
     #   jsonschema
     #   referencing
+certifi==2025.4.26
+    # via requests
+cfgv==3.4.0
+    # via pre-commit
+charset-normalizer==3.4.2
+    # via requests
 contourpy==1.3.2
     # via matplotlib
 cycler==0.12.1
@@ -24,6 +30,10 @@ deprecation==2.1.0
     # via
     #   fourcipp
     #   rapidyaml
+distlib==0.3.9
+    # via virtualenv
+filelock==3.18.0
+    # via virtualenv
 fonttools==4.57.0
     # via matplotlib
 fourcipp @ git+https://github.com/4C-multiphysics/fourcipp.git
@@ -32,8 +42,14 @@ frozenlist==1.6.0
     # via
     #   aiohttp
     #   aiosignal
+identify==2.6.10
+    # via pre-commit
 idna==3.10
-    # via yarl
+    # via
+    #   requests
+    #   yarl
+iniconfig==2.1.0
+    # via pytest
 jsonschema==4.23.0
     # via fourcipp
 jsonschema-rs==0.30.0
@@ -47,11 +63,15 @@ kiwisolver==1.4.8
 lnmmeshio==5.6.3
     # via -r requirements.in
 loguru==0.7.3
-    # via fourcipp
+    # via
+    #   fourcipp
+    #   lnmmeshio
 markdown-it-py==3.0.0
     # via rich
 matplotlib==3.10.1
-    # via vtk
+    # via
+    #   pyvista
+    #   vtk
 mdurl==0.1.2
     # via markdown-it-py
 meshio==5.3.5
@@ -66,6 +86,8 @@ multidict==6.4.3
     #   yarl
 narwhals==1.38.0
     # via plotly
+nodeenv==1.9.1
+    # via pre-commit
 numpy==2.2.5
     # via
     #   -r requirements.in
@@ -75,22 +97,35 @@ numpy==2.2.5
     #   matplotlib
     #   meshio
     #   pandas
+    #   pyvista
 packaging==25.0
     # via
     #   deprecation
     #   fourcipp
     #   matplotlib
     #   plotly
+    #   pooch
+    #   pytest
 pandas==2.2.3
     # via -r requirements.in
 pillow==11.2.1
-    # via matplotlib
+    # via
+    #   matplotlib
+    #   pyvista
+platformdirs==4.3.8
+    # via
+    #   pooch
+    #   virtualenv
 plotly==6.0.1
     # via
     #   -r requirements.in
     #   trame-plotly
-progress==1.6
-    # via lnmmeshio
+pluggy==1.6.0
+    # via pytest
+pooch==1.8.2
+    # via pyvista
+pre-commit==4.2.0
+    # via -r requirements.in
 propcache==0.3.1
     # via
     #   aiohttp
@@ -99,6 +134,8 @@ pygments==2.19.1
     # via rich
 pyparsing==3.2.3
     # via matplotlib
+pytest==8.3.5
+    # via -r requirements.in
 python-dateutil==2.9.0.post0
     # via
     #   matplotlib
@@ -112,12 +149,18 @@ pyvista==0.45.0
 pyyaml==6.0.2
     # via trame
 rapidyaml==0.9.0
+    # via
+    #   lnmmeshio
+    #   pre-commit
+    #   trame
     # via fourcipp
 referencing==0.36.2
     # via
     #   fourcipp
     #   jsonschema
     #   jsonschema-specifications
+requests==2.32.3
+    # via pooch
 rich==14.0.0
     # via meshio
 rpds-py==0.24.0
@@ -131,8 +174,12 @@ ruamel-yaml-clib==0.2.12
     # via
     #   fourcipp
     #   ruamel-yaml
+scooby==0.10.1
+    # via pyvista
 six==1.17.0
     # via python-dateutil
+tqdm==4.66.5
+    # via lnmmeshio
 trame==3.9.0
     # via -r requirements.in
 trame-client==3.8.2
@@ -160,11 +207,18 @@ typing-extensions==4.13.2
     # via
     #   fourcipp
     #   python-utils
+    #   pyvista
     #   referencing
 tzdata==2025.2
     # via pandas
+urllib3==2.4.0
+    # via requests
+virtualenv==20.31.2
+    # via pre-commit
 vtk==9.4.2
-    # via -r requirements.in
+    # via
+    #   -r requirements.in
+    #   pyvista
 wslink==2.3.3
     # via
     #   trame
